fix: resolve sign-compare and missing prototype warnings
- Use size_t for loop counters when comparing with sizeof()
- Use JDIMENSION type for JPEG image width comparisons
- Mark PNG callback functions as static with proper declarations
- Suppress unused parameter warnings with void casts

+++ src/image/oslLoadImageFile.c
@@ -12,7 +12,7 @@ OSL_IMAGE *oslLoadImageFile(char *filename, int location, int pixelFormat) {
 	}
 
 	char extension[10] = {0};
-	int i = 0;
+	size_t i = 0;
 	while (i < sizeof(extension) - 1 && ext[i]) // Leave space for the null terminator
 	{
 		extension[i] = tolower((unsigned char) ext[i]);
@@ -21,15 +21,15 @@ OSL_IMAGE *oslLoadImageFile(char *filename, int location, int pixelFormat) {
 
 	// Load the appropriate image type based on the extension
 #ifdef OSL_IMAGE_LOADER_PNG
-    if (strcmp(extension, ".png") == 0)
+	if (strcmp(extension, ".png") == 0)
 		return oslLoadImageFilePNG(filename, location, pixelFormat);
 #endif
 #ifdef OSL_IMAGE_LOADER_JPEG
-    if (strcmp(extension, ".jpg") == 0)
+	if (strcmp(extension, ".jpg") == 0)
 		return oslLoadImageFileJPG(filename, location, pixelFormat);
 #endif
 #ifdef OSL_IMAGE_LOADER_GIF
-    if (strcmp(extension, ".gif") == 0)
+	if (strcmp(extension, ".gif") == 0)
 		return oslLoadImageFileGIF(filename, location, pixelFormat);
 #endif
